Build preview images from `preview/*` branches
From time to time we need to build images with a feature preview to
test in an staging environment. This PR allows us to push a branch
named `preview/some-description` in order to build and push an image
with the `some-description` tag.

It's up to the user to check for name colisions - ie, if they push a
`preview/latest` branch.